+++ react-components/src/components/Architecture/SegmentedButtons.test.tsx
@@ -19,7 +19,7 @@ import {
   isEnabled,
   isSelected,
   isToggled
-} from '../../../tests/tests-utilities/cogs/cogsTestUtils';
+} from '../../../tests/tests-utilities/cogs/htmlTestUtils';
 
 describe(SegmentedButtons.name, () => {
   test('should render with default values', () => {
