+++ react-components/src/components/Architecture/SettingsButton.test.tsx
@@ -15,7 +15,7 @@ import { DividerCommand } from '../../architecture/base/commands/DividerCommand'
 import { TestSliderCommand } from '#test-utils/architecture/commands/TestSliderCommand';
 import { TestFilterCommand } from '#test-utils/architecture/commands/TestFilterCommand';
 import { TestOptionsCommand } from '#test-utils/architecture/commands/TestOptionsCommand';
-import { getButtonsInContainer } from '../../../tests/tests-utilities/cogs/cogsTestUtils';
+import { getButtonsInContainer } from '../../../tests/tests-utilities/cogs/htmlTestUtils';
 
 let wrapper: (props: PropsWithChildren) => ReactElement;
